PRACTICE-DONE/LABA-GRAPHQL/requirements.txt
fastapi[all]
strawberry-graphql[fastapi]
sqlalchemy[asyncio]
asyncpg
pydantic-settings
psycopg2-binary
uvicorn[standard]

